requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
ecdsa==0.19.0
enum-compat==0.0.3
gunicorn==22.0.0
idna==3.7
intuit-oauth==1.2.5
oauthlib==3.2.2
packaging==24.0
pyasn1==0.6.0
PyMySQL==1.1.0
python-jose==3.3.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
